requirements.txt
onnxruntime==1.16.2
opencv_python==4.8.1.78
Pillow==10.0.1
requests==2.31.0 
flask==3.0.0
numpy==1.26.2 
python-dotenv
selenium
pytest 
pytest-flask
webdriver-manager
pytest-cov 
psutil 
pillow